# Hustle-Safe/Hustle-Safe/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: HustleSafe API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workers:
    get:
      operationId: listWorkers
      tags: [workers]
      summary: List all workers
      parameters:
        - name: zone_id
          in: query
          schema:
            type: string
        - name: tier
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of workers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkerListResponse"
    post:
      operationId: createWorker
      tags: [workers]
      summary: Create a new worker
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkerRequest"
      responses:
        "201":
          description: Worker created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Worker"
  /workers/{id}:
    get:
      operationId: getWorker
      tags: [workers]
      summary: Get a worker by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Worker details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Worker"
        "404":
          description: Worker not found
  /workers/phone/{phone}:
    get:
      operationId: getWorkerByPhone
      tags: [workers]
      summary: Get a worker by phone number
      parameters:
        - name: phone
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Worker details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Worker"
        "404":
          description: Worker not found
...
  /disruptions:
    get:
      operationId: listDisruptions
      tags: [disruptions]
      summary: List disruption events
      parameters:
        - name: is_active
          in: query
          schema:
            type: boolean
        - name: zone_id
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of disruption events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DisruptionListResponse"
  /simulator/trigger:
    post:
      operationId: triggerDisruption
      tags: [simulator]
      summary: Trigger a disruption event in a zone
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TriggerDisruptionRequest"
      responses:
        "200":
          description: Disruption triggered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TriggerDisruptionResponse"
  /simulator/resolve/{zoneId}:
    post:
      operationId: resolveDisruption
      tags: [simulator]
      summary: Resolve/end a disruption in a zone
      parameters:
        - name: zoneId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Disruption resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResolveDisruptionResponse"
...
  /fraud/evaluate:
    post:
      operationId: evaluateFraudClaim
      tags: [fraud]
      summary: Evaluate fraud for a worker claim context using full adaptive pipeline
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FraudEvaluationRequest"
      responses:
        "200":
          description: Final Adaptive Fraud Evaluation Result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FraudEvaluationResult"
...
components:
  schemas:
...
    Worker:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        phone:
          type: string
        email:
          type: string
          nullable: true
        platform:
          type: string
          enum: [zomato, swiggy, both]
        zone_id:
          type: string
        upi_id:
          type: string
          nullable: true
        platform_rating:
          type: number
        policy_tier:
          type: string
          enum: [basic, standard, pro]
        is_active:
          type: boolean
        fraud_score:
          type: number
        account_age_days:
          type: integer
        created_at:
          type: string
      required:
        - id
        - name
        - phone
        - platform
        - zone_id
        - platform_rating
        - policy_tier
        - is_active
        - fraud_score
        - account_age_days
        - created_at
    WorkerListResponse:
      type: object
      properties:
        workers:
          type: array
          items:
            $ref: "#/components/schemas/Worker"
        total:
          type: integer
      required:
        - workers
        - total
    CreateWorkerRequest:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        platform:
          type: string
          enum: [zomato, swiggy, both]
        zone_id:
          type: string
        upi_id:
          type: string
        policy_tier:
          type: string
          enum: [basic, standard, pro]
      required:
        - name
        - phone
        - platform
        - zone_id
        - policy_tier
    Zone:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        city:
          type: string
        lat:
          type: number
          nullable: true
        lng:
          type: number
          nullable: true
        gds_score:
          type: integer
        status:
          type: string
          enum: [normal, elevated, high, shutdown]
        active_workers:
          type: integer
        rainfall_mm:
          type: number
        traffic_score:
          type: number
        aqi:
          type: integer
        demand_drop_pct:
          type: integer
        govt_alert:
          type: boolean
        last_updated:
          type: string
      required:
        - id
        - name
        - city
        - gds_score
        - status
        - active_workers
        - rainfall_mm
        - traffic_score
        - aqi
        - demand_drop_pct
        - govt_alert
        - last_updated
...
    DisruptionEvent:
      type: object
      properties:
        id:
          type: string
        zone_id:
          type: string
          nullable: true
        event_type:
          type: string
          enum: [heavy_rain, flood, curfew, platform_outage, strike, aqi_hazard]
        gds_target:
          type: integer
          nullable: true
        duration_minutes:
          type: integer
          nullable: true
        triggered_by:
          type: string
        is_active:
          type: boolean
        started_at:
          type: string
          nullable: true
        ended_at:
          type: string
          nullable: true
        created_at:
          type: string
      required:
        - id
        - event_type
        - triggered_by
        - is_active
        - created_at
    DisruptionListResponse:
      type: object
      properties:
        disruptions:
          type: array
          items:
            $ref: "#/components/schemas/DisruptionEvent"
      required:
        - disruptions
    TriggerDisruptionRequest:
      type: object
      properties:
        zone_id:
          type: string
        event_type:
          type: string
          enum: [heavy_rain, flood, curfew, platform_outage, strike, aqi_hazard]
        gds_target:
          type: integer
        duration_minutes:
          type: integer
      required:
        - zone_id
        - event_type
        - gds_target
    TriggerDisruptionResponse:
      type: object
      properties:
        success:
          type: boolean
        disruption:
          $ref: "#/components/schemas/DisruptionEvent"
        zone:
          $ref: "#/components/schemas/Zone"
        claims_created:
          type: integer
        message:
          type: string
      required:
        - success
        - disruption
        - zone
        - claims_created
        - message
    ResolveDisruptionResponse:
      type: object
      properties:
        success:
          type: boolean
        zone:
          $ref: "#/components/schemas/Zone"
        message:
          type: string
      required:
        - success
        - zone
        - message
...
    FraudEvaluationRequest:
      type: object
      required:
        - worker_id
        - zone_id
        - user_lat
        - user_lng
        - device_id
      properties:
        worker_id:
          type: string
          format: uuid
          description: Unique identifier of the worker.
        zone_id:
          type: string
          description: Identifier of the zone the worker claims to be disrupted in.
        user_lat:
          type: number
          format: float
          description: Worker's current latitude.
        user_lng:
          type: number
          format: float
          description: Worker's current longitude.
        device_id:
          type: string
          description: Fingerprint ID of the user's mobile device for fraud checks.
...
    FraudEvaluationResult:
      type: object
      required:
        - final_score
        - resolution_tier
        - individual_signal_scores
        - signal_status
        - critical_flags_fired
        - critical_flag_signals
        - worker_trust_score
        - adaptive_threshold_used
        - primary_elevated_signal
        - pass2_triggered
        - pass2_trigger_reason
        - adversarial_elevated_count
        - data_quality
      properties:
        final_score:
          type: number
          format: float
        resolution_tier:
          type: string
        individual_signal_scores:
          type: object
          additionalProperties:
            type: number
            nullable: true
        signal_status:
          type: object
          additionalProperties:
            type: string
            enum: [computed, missing_data, error]
        critical_flags_fired:
          type: boolean
        critical_flag_signals:
          type: array
          items:
            type: string
        worker_trust_score:
          type: number
          format: float
        adaptive_threshold_used:
          type: number
          format: float
        primary_elevated_signal:
          type: string
        pass2_triggered:
          type: boolean
        pass2_trigger_reason:
          type: string
          enum: [pass1_clear, pass1_amber, pass1_high_risk]
        adversarial_elevated_count:
          type: integer
          format: int32
        data_quality:
          type: object
          required:
            - available_signals
            - total_signals
            - availability_ratio
          properties:
            available_signals:
              type: integer
              format: int32
            total_signals:
              type: integer
              format: int32
            availability_ratio:
              type: number
              format: float
...