# Hustle-Safe/Hustle-Safe/lib/api-zod/src/generated/types/disruptionListResponse.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Api
 * HustleSafe API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DisruptionEvent } from "./disruptionEvent";

export interface DisruptionListResponse {
  disruptions: DisruptionEvent[];
}